actual-code/17-django/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
requests>=2.25,<3.0